requirements.txt
Django==1.7b2
django-extensions==1.3.3
djangorestframework==2.3.13
gnureadline==6.3.3
ipython==2.0.0
six==1.6.1
solrpy==0.9.6
wsgiref==0.1.2
